Added DMA and semaphores (not working presently)
If you comment out the semaphore usages in each UART thread, and use blocking I/O in the datawriter, then the robot moves perfectly. This is our "control" in the experiment that is getting this robot to work

+++ Robot/Inc/UART_Handler.h
@@ -24,15 +24,15 @@
 typedef struct {
 	uint8_t 					type;
 	Dynamixel_HandleTypeDef*	motorHandle;
-	uint16_t 					position;
-	uint16_t 					velocity;
+	float 						position;
+	float 						velocity;
 	QueueHandle_t				qHandle;
 }UARTcmd;
 
 typedef struct {
 	Dynamixel_HandleTypeDef*	motorHandle;
-	uint16_t 					position;
-	uint16_t 					velocity;
+	float 						position;
+	float 						velocity;
 }UARTrx;
 /* USER CODE END Private defines */
